Update README.md
Broovs (BRS) Projects is a software and technology development company established by Turkish Academics at the Technopolis Technology Development Center. Their projects have an ecosystem focused on benefits to people and society. At this point, 12 projects have been produced and their planning has been completed. 
We have noticed that all projects support each other and create a consumption mechanism in each project, plan a fiction on the BRS Token owned by the company, and since the fiction and the structure of the company have been researched and passed the level of sharing with you, this report has been prepared for your information purposes. The company plans to make the BRS TOKEN, which is produced with high ease of use and awareness on the BSC network, an asset that will be used in many projects in the future. The company solves its projects with the support of its in-house technical team and academicians without the need for external procurement, and within the scope of the project,
1.	The dec generation search engine Broovs.com the beta version is active and has a user-oriented algorithm whose entries are made and approved by users. https://broovs.com 
2.	Playsore now active in the messenger app with Swift, now that are not available in any messaging app crypto wallet feature added that it belongs to any contract at the time of the BSc network to send and receive coins, it is possible https://swiftmessengerweb.com   
3.	The company has also identified a shortage in Turkey within the scope of BRS Projects and has been able to successfully list the BSC network https://coinswifter.com CoinSwifter, registered at the address, has activated the crypto trading platform. With this platform, easy and fast transactions can be made without using any brokerage exchange. 
4.	Simultaneously with the installation of the exchange, the secure Cold Wallet Swifter Wallet has been activated with the ability to carry out swap transactions and high-security storage. https://swifterwallet.com
5.	As a result of long-term studies, the Company has activated the XRAY-RAYCHAIN Blockchain, which they describe as the first open source Blockchain in Turkey, and the premining phase is underway. https://raychainscan.com 
6.	Turkey's first and only real Xray blockchain ’s NFT produce to be able to broovsblock broovsbox smart contract to be able to the next company and doing projects with companies, institutions and individuals of data security, data recording and data storage node participating in the network to perform operations that holds and each point will become a supporter of an xray. With the application that will be installed with the algorithm developed by our company based on the XRAY Blockchain, the memory of mobile phones will be increased up to 1 TB, which will prevent people from having to constantly change their mobile phones and ensure that the capital will remain in our country.
7.	The development stage is ongoing Fanmeter.tv with its project, it aims to achieve a first in the world by creating a scoring system for social media phenomena and ensuring that advertisers and families reach the right audience. https://fanmeter.tv 
We also mentioned Broovs.com , Swift Messenger, and Fanmeter.tv their projects are user-oriented projects and are designed as Web 3.0 projects that share their income with users after the developer shares have been separated and do not keep personal records. To get more information https://broovs.io you can visit his address.